añadido el parametro cantidad en el metodo operativa_cuenta: cantidad

+++ ProyectoETSOwen/src/main/java/cuentas/CCuenta.java
@@ -40,7 +40,7 @@ public class CCuenta {
         saldo = saldo - cantidad;
     }
     
-    public void operativa_cuenta(String accion, double cantidad) throws Exception{
+    public void operativa_cuenta(String accion, float cantidad) throws Exception{
         if (accion.toLowerCase().equals("retirar")){
             this.retirar(cantidad);
         } else if (accion.toLowerCase().equals("ingresar")) {
